+++ App/Client/ViewSplitter.h
@@ -3,8 +3,10 @@
 #ifndef VIEWSPLITTER_H
 #define VIEWSPLITTER_H
 
-#include "View.h"
 #include <QSplitter>
+#include "ViewSplitterContainer.h"
+#include "View.h"
+
 /*!
  * \brief The split view class
  * \ingroup ViewApi
@@ -14,7 +16,7 @@ class CViewSplitter : public CView
     Q_OBJECT
 
 public:
-    CViewSplitter(QWidget *parent = nullptr);
+    CViewSplitter(CParameterApp* pPara, QWidget *parent = nullptr);
     virtual ~CViewSplitter();
 
     // CView interface
@@ -29,17 +31,21 @@ public:
     virtual int SetFullScreen(bool bFull) override;
 
 private:
-    int m_nRow;
-    int m_nCount;
-    int m_nIdxRow, m_nIdxCol;
+    int m_nRow;   // Row number
+    int m_nCount; // Total number
+    int m_nIdxRow, m_nIdxCol;  // Index
+
     QSplitter* m_pMain;
     QVector<QSplitter*> m_Row;
+
+    QMap<QWidget*, CViewSplitterContainer*> m_Container;
     QString m_szStyleSheet;
     int m_HandleWidth;
 
 private:
     int GetIndex(QWidget* pView, int &nRow, int &nCol);
-    QWidget* GetView(const int &nRow, const int &nCol);
+    CViewSplitterContainer* GetContainer(const int &nRow, const int &nCol);
+    CViewSplitterContainer* GetContainer(QWidget* pView);
 };
 
 #endif // VIEWSPLITTER_H